Fix services sql

+++ ew/src/workers/sigmausd/store/schema.sql
@@ -79,7 +79,7 @@ create table sigmausd.history (
 */
 create table sigmausd.services (
     -- address id of service or null for direct interaction
-    address_id integer primary key,
+    address_id bigint primary key,
     -- total transactions to date
     tx_count bigint not null,
     -- first and last tx timestamps
